requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.6.15
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-mail==1.1.4
greenlet==1.1.2
h11==0.13.0
httpcore==0.13.2
httpx==0.18.1
idna==3.3
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyasn1==0.4.8
pydantic==1.9.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
flower
redis
pytest~=7.1.2